pyproject.toml
[project]
name = "sql-crew"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "crewai[google-genai]>=1.8.0",
    "crewai-tools>=0.1.0",
    "groq>=0.37.1",
    "ipython>=9.9.0",
    "langchain-anthropic>=0.3.22",
    "langchain-core>=0.3.83",
    "langchain-google-genai>=2.0.8",
    "langchain-groq>=0.3.8",
    "langchain-openai>=0.3.23",
    "langgraph>=1.0.1",
    "mysql-connector-python>=9.5.0",
    "pandas>=2.3.3",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.1.1",
    "rapidfuzz>=3.14.3",
    "sqlalchemy>=2.0.45",
    "tqdm>=4.67.1",
]
